requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
python-dateutil==2.8.2
requests==2.31.0
stripe==7.8.0
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0
freezegun==1.2.2
responses==0.24.1
